+++ app/src/main/java/com/github/members/directory/di/RepositoryModule.kt
@@ -1,6 +1,6 @@
 package com.github.members.directory.di
 
-import com.github.members.directory.features.git_user.repository.PagingRepositoryMembers
+import com.github.members.directory.features.users.repository.PagingRepositoryMembers
 import org.koin.dsl.module
 
 val repositoryModule = module {
