+++ app/src/main/java/com/github/members/directory/di/ViewModelModule.kt
@@ -1,6 +1,6 @@
 package com.github.members.directory.di
 
-import com.github.members.directory.features.git_user.MembersViewModel
+import com.github.members.directory.features.users.MembersViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
