+++ app/src/main/java/com/github/members/directory/features/users/MembersViewModel.kt
@@ -1,4 +1,4 @@
-package com.github.members.directory.features.git_user
+package com.github.members.directory.features.users
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,7 +6,7 @@ import androidx.paging.ExperimentalPagingApi
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.github.members.baseplate_persistence.model.DBMembers
-import com.github.members.directory.features.git_user.repository.PagingRepositoryMembers
+import com.github.members.directory.features.users.repository.PagingRepositoryMembers
 import kotlinx.coroutines.flow.Flow
