+++ app/src/main/java/com/github/members/directory/features/users/adapter/MembersPagingViewHolder.kt
@@ -1,4 +1,4 @@
-package com.github.members.directory.features.git_user.adapter
+package com.github.members.directory.features.users.adapter
 
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
