+++ app/src/main/java/com/github/members/directory/features/users/repository/PagingRepositoryMembers.kt
@@ -1,4 +1,4 @@
-package com.github.members.directory.features.git_user.repository
+package com.github.members.directory.features.users.repository
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.Pager
